requirements.txt
altair==5.5.0
attrs==23.2.0
blinker==1.9.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
narwhals==1.33.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
pyarrow==15.0.2
pydeck==0.8.1b0
python-dateutil==2.8.2
pytz==2024.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.15.2
six==1.16.0
smmap==5.0.1
streamlit==1.31.1
tenacity==8.2.3
toml==0.10.2
tornado==6.3.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
torch
torchvision
